+++ SecureChatServer.h
@@ -21,7 +21,7 @@ class SecureChatServer{
 
         //Port and listening IP address, in dotted notation (e.g. 192.168.1.1)
         char address[16];
-        uint16_t port;
+        int port;
         int listening_socket;
         struct sockaddr_in server_addr;
 
@@ -66,5 +66,5 @@ class SecureChatServer{
 
     public:
         //Constructor that gets as inputs the address, the port and the user filename.
-        SecureChatServer(const char* addr, uint16_t port, const char *user_filename);
+        SecureChatServer(const char* addr, int port, const char *user_filename);
 };
